ADD// fix du collapse + add margin page 404 + light reorganize des sous-dossiers
