+++ src/components/Error/_error404.scss
@@ -3,6 +3,7 @@
 .error404 {
     @include flex($direction:column,$align:center,$justify:normal,$gap:113px);
     @include font($font:$global-font,$weight: bold,$color:$red);
+    margin-top: 60px;
 
     .error404-title {
         font-size: 96px;
@@ -11,13 +12,14 @@
 
     .error404-text {
         font-size: 18px;
-        font-weight: normal;
-        width: 68%;
+        font-weight: 500;
+        width: 70%;
+        text-align: center;
     }
 
     .error404-homelink {
         font-size: 14px;
-        font-weight: normal;
+        font-weight: 500;
         color: $black;
     }
 }
