+++ src/components/Main/Card/Carousel/_carousel.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/sass/utils/styling';
+@import '../../../../styles/sass/utils/styling';
 
 .carousel {
 	@include flex($direction:column,$align:center,$justify:center,$gap:unset);
