+++ src/components/Main/Card/Rating/Rating.js
