+++ src/components/Main/Card/_card.scss
@@ -24,7 +24,6 @@
 	.card-tags {
 		@include font($font: $global-font,$size: 10px, $weight:bold,$color:$white);
 		@include flex($direction:row,$align:center,$justify:flex-start,$gap:10px);
-		// @include margin($margin:10px 0,$padding:unset);
 		flex-wrap: wrap;
 
 		.card-tag {
@@ -71,44 +70,6 @@
 
 	.card-details {
 		@include flex($direction:column,$align:normal,$justify:normal,$gap:15px);
-
-		.collapse {
-			@include font($font:$global-font,$size:13px,$weight:normal,$color:$black);
-
-			&-btn {
-				@include flex($direction:row-reverse,$align:center,$justify:space-between,$gap:unset);
-				@include font($font:$global-font,$size:13px,$weight:bold,$color:$white);
-				border: none;
-				border-radius: 5px;
-				background-color: $red;
-				width: 100%;
-				height: 30px;
-				text-align: start;
-				padding: 5px 10px;
-			}
-
-			.collapse-icon {
-				filter: invert(1);
-				transition: transform 0.5s ease;
-
-				&.open {
-					transform: rotate(-90deg);
-				}
-				  
-				&.closed {
-					transform: rotate(90deg);
-				}
-			}
-
-			.collapse-description {
-				padding: 10px 15px;
-				transition: transform 0.5s ease;
-			
-				&.open {
-					
-				}
-			}
-		}
 	}
 }
